Move selects to another file
