+++ prompts/templates/models.go
@@ -4,11 +4,8 @@ import (
 	"sort"
 	"strings"
 
-	manifold "github.com/manifoldco/go-manifold"
-
 	cModels "github.com/manifoldco/manifold-cli/generated/catalog/models"
 	iModels "github.com/manifoldco/manifold-cli/generated/identity/models"
-
 	mModels "github.com/manifoldco/manifold-cli/generated/marketplace/models"
 )
 
@@ -18,14 +15,14 @@ type Provider struct {
 }
 
 type Resource struct {
-	Name    manifold.Label
-	Title   manifold.Name
-	Project manifold.Label
+	Name    string
+	Title   string
+	Project string
 }
 
 type Project struct {
-	Name  manifold.Label
-	Title manifold.Name
+	Name  string
+	Title string
 }
 
 type Team struct {
@@ -38,14 +35,14 @@ func Resources(list []*mModels.Resource, projects []*mModels.Project) []Resource
 
 	for i, m := range list {
 		r := Resource{
-			Name:  m.Body.Label,
-			Title: m.Body.Name,
+			Name:  string(m.Body.Label),
+			Title: string(m.Body.Name),
 		}
 
 		if m.Body.ProjectID != nil {
 			for _, p := range projects {
 				if *m.Body.ProjectID == p.ID {
-					r.Project = p.Body.Label
+					r.Project = string(p.Body.Label)
 				}
 			}
 		}
@@ -66,11 +63,11 @@ func Providers(list []*cModels.Provider) []Provider {
 	providers := make([]Provider, len(list))
 
 	for i, m := range list {
-		t := Provider{
+		p := Provider{
 			Name:  string(m.Body.Label),
 			Title: string(m.Body.Name),
 		}
-		providers[i] = t
+		providers[i] = p
 	}
 
 	sort.Slice(providers, func(i, j int) bool {
@@ -100,4 +97,24 @@ func Teams(list []*iModels.Team) []Team {
 	})
 
 	return teams
+}
+
+func Projects(list []*mModels.Project) []Project {
+	projects := make([]Project, len(list))
+
+	for i, m := range list {
+		p := Project{
+			Name:  string(m.Body.Label),
+			Title: string(m.Body.Name),
+		}
+		projects[i] = p
+	}
+
+	sort.Slice(projects, func(i, j int) bool {
+		a := string(projects[i].Name)
+		b := string(projects[j].Name)
+		return strings.ToLower(a) < strings.ToLower(b)
+	})
+
+	return projects
 }
